+++ spec/osm/badges_spec.rb
@@ -75,11 +75,39 @@ describe "Badges" do
         }
       }
       FakeWeb.register_uri(:post, "https://www.onlinescoutmanager.co.uk/challenges.php?action=getInitialBadges&type=core&sectionid=1&section=beavers&termid=2", :body => badges_body.to_json)
+      Osm::Term.stub(:get_current_term_for_section) { Osm::Term.new(:id => 2) }
 
       section = Osm::Section.new(:id => 1, :type => :beavers)
-      section.should_not be_nil
-      Osm::Badges.get_stock(@api, section, 2).should == {'badge_1' => 1, 'badge_2' => 2}
+      Osm::Badges.get_stock(@api, section).should == {'badge_1' => 1, 'badge_2' => 2}
     end
+
+    describe "Update badge stock levels" do
+
+      it "Succeds" do
+        url = "https://www.onlinescoutmanager.co.uk/challenges.php?action=updateStock"
+        HTTParty.should_receive(:post).with(url, {:body => {
+          'apiid' => @CONFIGURATION[:api][:osm][:id],
+          'token' => @CONFIGURATION[:api][:osm][:token],
+          'userid' => 'user_id',
+          'secret' => 'secret',
+          'stock' => 10,
+          'table' => 'badge',
+          'sectionid' => 1,
+          'section' => :beavers,
+        }}) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"sectionid":"1","badge":"10"}'}) }
+
+        section = Osm::Section.new(:id => 1, :type => :beavers)
+        Osm::Badges.update_stock(@api, section, 'badge', 10).should be_true
+      end
+
+      it "Fails" do
+        HTTParty.stub(:post) { OsmTest::DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"sectionid":"1","badge":"1"}'}) }
+        section = Osm::Section.new(:id => 1, :type => :beavers)
+        Osm::Badges.update_stock(@api, section, 'badge', 10).should be_false
+      end
+
+    end
+
   end
 
 end